feat: remove extension in favor of native-supported partial errors in execute
chore: remove extra md file I meant to remove in last commit
