+++ docs/guides/errors.mdx
@@ -163,8 +163,14 @@ schema = strawberry.Schema(query=Query)
 
 ## Expected errors
 
-If an error is expected then it is often best to express it in the schema. This
-allows the client to deal with the error in a robust way.
+Handling expected errors is typically approached in one or both of two ways:
+
+1. Return the error as part of the return type
+2. Append the error to the `errors` array
+
+### Returning as part of the return type
+
+It is often best to express expected errors in the schema. This allows the client to deal with the error in a robust way.
 
 This could be achieved by making the field optional when there is a possibility
 that the data won’t exist:
@@ -251,6 +257,82 @@ mutation RegisterUser($username: String!, $password: String!) {
 This approach allows you to express the possible error states in the schema and
 so provide a robust interface for your client to account for all the potential
 outcomes from a mutation.
+
+### Appending to `errors` array
+
+As seen in the [Unhandled execution errors][#unhandled-execution-errors] section above, errors are automatically added to the errors array by raising the error in execution.
+
+But cases may arise where you want to handle errors or otherwise use the errors array to indicate that issues have occurred, but still return data in the response too.
+
+This can be achieved by adding `Exception` instances to the `partial_errors` array on the context.
+
+```python
+import strawberry
+
+
+@strawberry.field
+def query(self, info) -> bool:
+    info.context.partial_errors.append(Exception("Partial failure"))
+    info.context.partial_errors.append(MyCustomException("Another partial failure"))
+    return True
+```
+
+Results:
+
+```json
+{
+  "data": {
+    "query": true
+  },
+  "errors": [
+    {
+      "message": "Partial failure"
+    }
+  ]
+}
+```
+
+Note that when exceptions are raised in resolvers, the GraphQL library catches the exception and maps it to a `GraphQLError` with path and location data pulled from the resolver.
+If you'd like to persist this kind of metadata in your response, you'll need to add a `GraphQLError` with that data yourself, commonly accomplished with the `located_error()` function.
+
+```python
+import strawberry
+
+from graphql import located_error
+
+
+@strawberry.field
+def query(self, info) -> bool:
+    nodes = [next(n for n in info.field_nodes if n.name.value == "query")]
+    info.context.partial_errors.append(
+        located_error(
+            Exception("Error with location and path information"),
+            nodes=nodes,
+            path=info.path.as_list(),
+        ),
+    )
+    return True
+```
+
+Results:
+
+```json
+{
+  "data": {
+    "query": true
+  },
+  "errors": [
+    {
+      "message": "Error with location and path information",
+      "location": {
+        "line": 1,
+        "column": 9
+      },
+      "path": ["query"]
+    }
+  ]
+}
+```
 
 ---
