+++ strawberry/extensions/__init__.py
@@ -9,7 +9,6 @@ from .mask_errors import MaskErrors
 from .max_aliases import MaxAliasesLimiter
 from .max_tokens import MaxTokensLimiter
 from .parser_cache import ParserCache
-from .partial_results import PartialResultsExtension
 from .query_depth_limiter import IgnoreContext, QueryDepthLimiter
 from .validation_cache import ValidationCache
 
@@ -36,7 +35,6 @@ __all__ = [
     "AddValidationRules",
     "DisableValidation",
     "ParserCache",
-    "PartialResultsExtension",
     "QueryDepthLimiter",
     "IgnoreContext",
     "ValidationCache",
